refactor(oxfmt): Rename cli/format -> cli/runner (#21407)
Just rename.
