+++ apps/oxfmt/src/cli/mod.rs
@@ -1,8 +1,8 @@
 pub(crate) mod command;
-mod format;
 mod init;
 mod reporter;
 mod result;
+mod runner;
 mod service;
 mod walk;
 
@@ -10,6 +10,6 @@ pub use crate::core::utils::init_tracing;
 #[cfg(feature = "napi")]
 pub use command::MigrateSource;
 pub use command::{FormatCommand, Mode, format_command};
-pub use format::CliRunner;
 pub use init::{init_miette, init_rayon};
 pub use result::CliRunResult;
+pub use runner::CliRunner;
